requirements.txt
robotframework==4.0.3
-e ./libraries/
